[#988] ROBOT axioms command (remove/annotate)
