--- robot-command/src/main/java/org/obolibrary/robot/RemoveAxiomsCommand.java
@@ -1,110 +1,0 @@
-package org.obolibrary.robot;
-
-import org.apache.commons.cli.CommandLine;
-import org.apache.commons.cli.Options;
-import org.semanticweb.owlapi.model.OWLOntology;
-import org.semanticweb.owlapi.model.OWLOntologyManager;
-
-/**
- * Remove axioms from an ontology based on a series of inputs.
- *
- * @author <a href="mailto:rctauber@gmail.com">Becky Tauber</a>
- */
-public class RemoveAxiomsCommand implements Command {
-
-  /** Store the command-line options for the command. */
-  private Options options;
-
-  /** Initialze the command. */
-  public RemoveAxiomsCommand() {
-    Options o = CommandLineHelper.getCommonOptions();
-    o.addOption("i", "input", true, "load ontology from a file");
-    o.addOption("I", "input-iri", true, "load ontology from an IRI");
-    o.addOption("o", "output", true, "save ontology to a file");
-    o.addOption(null, "base-iri", true, "specify a base namespace");
-    o.addOption("s", "select", true, "select a set of terms based on relations");
-    options = o;
-  }
-
-  /**
-   * Name of the command.
-   *
-   * @return name
-   */
-  public String getName() {
-    return "remove-axioms";
-  }
-
-  /**
-   * Brief description of the command.
-   *
-   * @return description
-   */
-  public String getDescription() {
-    return "remove deprecated axioms from an ontology";
-  }
-
-  /**
-   * Command-line usage for the command.
-   *
-   * @return usage
-   */
-  public String getUsage() {
-    return "robot " + getName() + " --input <file> " + "--output <file>";
-  }
-
-  /**
-   * Command-line options for the command.
-   *
-   * @return options
-   */
-  public Options getOptions() {
-    return options;
-  }
-
-  /**
-   * Handle the command-line and file operations.
-   *
-   * @param args strings to use as arguments
-   */
-  public void main(String[] args) {
-    try {
-      execute(null, args);
-    } catch (Exception e) {
-      CommandLineHelper.handleException(getUsage(), getOptions(), e);
-    }
-  }
-
-  /**
-   * Given an input state and command line arguments, create a new ontology with removed axioms and
-   * return a new state. The input ontology is not changed.
-   *
-   * @param state the state from the previous command, or null
-   * @param args the command-line arguments
-   * @return a new state with the new ontology
-   * @throws Exception on any problem
-   */
-  public CommandState execute(CommandState state, String[] args) throws Exception {
-    CommandLine line = CommandLineHelper.getCommandLine(getUsage(), getOptions(), args);
-    if (line == null) {
-      return null;
-    }
-
-    IOHelper ioHelper = CommandLineHelper.getIOHelper(line);
-    state = CommandLineHelper.updateInputOntology(ioHelper, state, line);
-    OWLOntology ontology = state.getOntology();
-    OWLOntologyManager manager = ontology.getOWLOntologyManager();
-
-    // Get a set of relation types, or annotations to select
-    final String select = CommandLineHelper.getOptionalValue(line, "select");
-
-    manager.removeAxioms(
-        ontology,
-        RelatedObjectsHelper.selectAxiomsByPattern(ontology, ioHelper, select));
-
-    // Save the changed ontology and return the state
-    CommandLineHelper.maybeSaveOutput(line, ontology);
-    state.setOntology(ontology);
-    return state;
-  }
-}
